♻️ 새소식 getpost 리팩토링
